% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assessment.R
\name{calculate_proportions}
\alias{calculate_proportions}
\title{Calculate proportions}
\usage{
calculate_proportions(data, plot = FALSE, prob_col = "prob.pos", bins = NULL)
}
\arguments{
\item{data}{A data frame (ideally, a tibble) with probabilities, in this case, of landslide initiation. Must be coercible to a data table.}

\item{plot}{Change this value to FALSE if you do not want to produce the success rate curve plot.}

\item{prob_col}{The name of the column which contains the probabilities.}

\item{bins}{A vector with breaks for binning the cumulative probability sums. This is useful when combining multiple success rate curves, which might occur if a larger basin is divided into smaller sub-basin DEMs. If you do not want the curve to be binned, set this argument to NULL.}
}
\value{
A \code{tibble} with the success rate curve.
}
\description{
This code takes a data frame of predicted landslide initiation values and calculates a success rate curve.
}
\details{
This calculation is done by ordering all observations in ascending order and getting cumulative sums for both the probabilities and the area.
These sums are then divided by the total values for probability and area, leaving us with cumulative proportion values.

This function support binning, which allow multiple success rate curves to be combined.

The difference between the success rate curve and the prediction rate curve is in the data that is used to generate it.
Success rate curves are generated using study areas.
Prediction rate curves are generated on entire prediction areas - or new data.
}
