% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assessment.R
\name{combine_proportions}
\alias{combine_proportions}
\title{Combine area and probability proportions}
\usage{
combine_proportions(..., plot = FALSE)
}
\arguments{
\item{...}{Any number of success rate curves to be combined.}

\item{plot}{Change to FALSE if you do not want to produce a plot showing the curves.}
}
\value{
A \code{tibble} with the combined success rate curve.
}
\description{
This function combines success rate curves into one curve. This is a weighted average, since the curves could represent different amounts of data.
The curves should be in the same format as produced by the calculate_proportions() function provided in this package and all must be the same length.
}
\details{
Calling this function with one curve returns the same curve with nothing changed.
}
