% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_apply.R
\name{\%col\%}
\alias{\%col\%}
\title{Apply on columns operator}
\usage{
x \%col\% y
}
\arguments{
\item{x}{matrix or data.frame.}

\item{y}{name of the function to apply on columns - you can define your own.}
}
\value{
Returns a vector or array or list of values obtained by applying the function on columns.
}
\description{
Apply on columns operator
}
\examples{
# can be used like:
iris[,c(1:4)] \%col\% function(x) mean(x, na.rm = T)

# or like this:
iris[,c(1:4)] \%col\% sum
}
