% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_op.R
\name{\%g\%}
\alias{\%g\%}
\title{Grepl operator}
\usage{
x \%g\% y
}
\arguments{
\item{x}{pattern: a character string we look for.}

\item{y}{a character vector where matches are sought.}
}
\value{
A logical vector.
}
\description{
Grepl operator
}
\examples{
# Returns TRUE
x <- "h"
y <- "hello"
x \%g\% y
}
