% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negate in operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector or NULL: the values to not be matched.}

\item{y}{vector or NULL: the values to not be matched against.}
}
\value{
A logical vector.
}
\description{
Negate in operator
}
\examples{
# Negates the \%in\% operator - returns c(FALSE, FALSE, FALSE)
x <- c("a", "b", "c")
x \%nin\% letters
}
