% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_op.R
\name{\%p\%}
\alias{\%p\%}
\title{paste0 operator with underscore separator}
\usage{
x \%p\% y
}
\arguments{
\item{x}{A character string.}

\item{y}{A character string.}
}
\value{
A character string.
}
\description{
paste0 operator with underscore separator
}
\examples{
# return "hello_world"
x <- "hello"
y <- "world"
x \%p\% y

# you can chain them - a_b_c
letters[1] \%p\% letters[2] \%p\% letters[3]
}
