% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_apply.R
\name{\%row\%}
\alias{\%row\%}
\title{Apply on rows operator}
\usage{
x \%row\% y
}
\arguments{
\item{x}{matrix or data.frame.}

\item{y}{name of the function to apply on rows - you can define your own.}
}
\value{
Returns a vector or array or list of values obtained by applying the function on rows.
}
\description{
Apply on rows operator
}
\examples{
# can be used like:
iris[,c(1:4)] \%row\% function(x) mean(x, na.rm = T)

# or like this:
iris[,c(1:4)] \%row\% sum
}
