% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dems.R
\name{predict_multiple_dems}
\alias{predict_multiple_dems}
\title{Predict multiple DEMs}
\usage{
predict_multiple_dems(model, files_in, output_dir, out_label, ...)
}
\arguments{
\item{model}{A trained model to be used for prediction. Must have function $predict_newdata and $state$train_task$feature_names, as provided by mlr3 models.}

\item{pred_dir}{A directory where the basin input data can be found.}

\item{basin_list}{A list of basins and sub-basins that correspond to the file structure in pred_dir.}

\item{out_dir}{An folder to write output data. If this is not specified, a new folder called \code{"predicting_output/"} will be created in the same parent directory as the predicting input directory.}

\item{scale_vals}{A tibble containing values to center and scale the data. These values should match what was used for the training data. It is the users responsibility to make sure they are transforming the data correctly according to their workflow.}

\item{output}{The type of file to write the predictions to. The default is a .csv file. Multiple types can be specified.}
}
\value{
Nothing
}
\description{
This function uses mapply() and predict_and_save to predict a full list of DEMs.
The predicting directory must contain a folder/file structure that follows the basins in the basin list. Each sub-directory should contain gradient, mean_curv, and pca files.
}
\details{
Output is written to files in the same file structure.
}
