% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assessment.R
\name{success_rate_auc}
\alias{success_rate_auc}
\title{Area under the success rate curve}
\usage{
success_rate_auc(curve, integral_type = "trap")
}
\arguments{
\item{curve}{The success rate curve. It should have a column "area_prop" and a column "prob_prop"}

\item{integral_type}{How the integral is calculated. Can be left, right, or trap.}
}
\value{
The area under the curve.
}
\description{
Calculates the area under the success rate curve as a single value.
This is a useful quantification of the success rate curve metric for easily comparing models and integrating this with a modeling ecosystem.
}
