% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assessment.R
\name{success_rate_curve}
\alias{success_rate_curve}
\title{Success Rate Curve}
\usage{
success_rate_curve(
  obs,
  modeled,
  plot = TRUE,
  prob_col = "prob.pos",
  bins = NULL,
  quiet = TRUE
)
}
\arguments{
\item{plot}{Whether to produce a default plot of the success rate curve and the modeled probability proportions.}

\item{prob_col}{The name of the predicted probability column. Default is "prob.pos"}

\item{bins}{}

\item{landslides}{A data frame of every landslide initiation point, with the predicted probability column.}

\item{dem_list}{A list of .csv files with predicted probabilities for each DEM.}
}
\value{
A data frame with the success rate curve and the modeled probability curve.
}
\description{
This function builds a success rate curve using landslide data and a list of the DEMs. A prediction rate curve can be produced if the landslides given have not been included in the model used for predicted probabilities (test data).
It is up to the user to ensure that all relevant DEMs are included in the list - the extent of the study area will impact the curve.
}
\details{
The DEM list should be filenames that lead to .csv files with predictions for each DEM. The .csv files should be in a format similar to what is produced in a accompanying function, with a probability output column from a model.
}
