% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dems.R
\name{tif_to_flt}
\alias{tif_to_flt}
\title{Tif to flt file conversion}
\usage{
tif_to_flt(
  path,
  new_path = paste0(gsub("\\\\..*", "", path), ".hdr"),
  rm_old = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{path}{Path of the file to convert}

\item{new_path}{The path to of where to write the file. The directory must
exist already. The default is a file of the same name as the
original file.}

\item{rm_old}{Whether to delete the .tif (and associated) files.}

\item{quiet}{Whether to report the time spent reading + writing files.}
}
\description{
This is a helpful tool for converting between Flt (ArcGIS raster files) and
GTiff files. The FORTRAN executables included in the TerrainWorks R packages
currently use .flt files, but .tif files are generally a better format for
storage, since they support compression.
}
\details{
This function offers an option to delete the .tif file, once it has been
converted. This option is included for space efficiency, if working with
large amounts of data.
}
